Separated draw of layered pane from JFrame, only for the demo is a Frame used, otherwise all other classes extend from JPanel
