+++ src/com/dis2/cards/simpleCard.java
@@ -10,6 +10,7 @@ import com.dis2.shared.Palette;
 
 
 public class simpleCard extends JPanel{
+	private static final long serialVersionUID = 1L;
 	
 	public Palette palette = new Palette();
 	cardWidget cw;
